pyproject.toml
[tool.poetry]
name = "mcp-customer-service"
version = "0.1.0"
description = "MCP Customer Service Assistant - Building AI Integrations with the Model Context Protocol"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastmcp = "^2.8.0"
pydantic = "^2.11.5"
python-dotenv = "^1.0.0"
openai = "^1.86.0"
anthropic = "^0.54.0"
langchain = "^0.3.0"
langchain-openai = "^0.3.22"
langchain-mcp-adapters = "^0.1.7"
dspy-ai = "^2.6.27"
litellm = "^1.72.4"
asyncio = "^3.4.3"
langchain-core = "^0.3.65"
langsmith = "^0.3.45"
langgraph = "^0.4.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
